services/ingestion-worker/pyproject.toml
[tool.poetry]
name = "ingestion-worker"
version = "0.1.0"
description = "PREDATOR Analytics Ingestion Worker"
authors = ["Dima <dima@example.com>"]

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.28"}
asyncpg = "^0.29.0"
neo4j = "^5.18.0"
aiokafka = "^0.10.0"
pydantic-settings = "^2.2.1"
predator-common = {path = "../../libs/predator-common", develop = true}
structlog = "^24.1.0"
setuptools = "^69.2.0"
minio = "^7.2.5"
httpx = "^0.27.0"
chardet = "^5.2.0"
aiohttp = "^3.9.0"
sentence-transformers = {version = "^2.6.0", optional = true}
qdrant-client = {version = "^1.8.0", optional = true}

[tool.poetry.extras]
ml = ["sentence-transformers", "qdrant-client"]

[tool.poetry.dev-dependencies]
pytest = "^8.1.1"
pytest-asyncio = "^0.23.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
ignore_missing_imports = true
strict = true
